Refactor project setup to modularize prompts, styles, and project structure; enhance CLI tool for creating Vite + React apps with CSS framework support and optional packages.
