+++ src/templates/index.js
@@ -1,0 +1,2 @@
+export { createAppTemplate } from "./app.js";
+export { createMainTemplate } from "./main.js";
